[test] wiring/no_fixture: move some OOM stress tests into wiring/no_fixture_stress as no_fixture overflows on Electron with test runner framework enabled
