+++ user/tests/wiring/no_fixture_stress/no_fixture_stress.spec.js
@@ -1,3 +1,3 @@
 suite('No fixture stress');
 
-platform('gen3');
+platform('gen3', 'gen2');
